refactor(refactor the folder structure): refactor the folder structure
