+++ src/Models/Item.ts
@@ -1,6 +1,6 @@
 import { Event } from './Event';
 
 export type Item = {
-    title: string;
-    events: Event[];
+  title: string;
+  events: Event[];
 };
