+++ src/index.tsx
@@ -1,2 +1,2 @@
-import Timeline from './Timeline';
+import Timeline from './Components/timeline/Timeline';
 export default Timeline;
